requirements.txt
pandas
openpyxl
sentence-transformers
python-jobspy